+++ modules/catalog/src/main/scala/lucuma/catalog/telluric/TelluricTargetsClient.scala
@@ -4,16 +4,17 @@
 package lucuma.catalog.telluric
 
 import cats.Applicative
-import cats.ApplicativeThrow
 import cats.effect.Concurrent
 import cats.syntax.all.*
 import clue.http4s.Http4sHttpBackend
 import clue.http4s.Http4sHttpClient
-import io.circe.syntax.*
+import clue.FetchClient
+import clue.syntax.*
 import org.http4s.Uri
 import org.http4s.client.Client
 import org.typelevel.log4cats.Logger
-import org.typelevel.log4cats.syntax.*
+import lucuma.catalog.telluric.TelluricSearchQuery.TelluricStar
+import lucuma.core.model.TelluricType
 
 /**
  * Client for the telluric targets service
@@ -31,46 +32,25 @@ object TelluricTargetsClient:
     Http4sHttpClient
       .of[F, TelluricService](uri)
       .map: http =>
+        given FetchClient[F, TelluricService] = http
+
         new TelluricTargetsClient[F]:
           override def search(input: TelluricSearchInput): F[List[TelluricStar]] =
-            val encodedVars = TelluricSearchQuery.varEncoder(input).asJson
-
-            for {
-              _        <- debug"Telluric targets call: $uri ${encodedVars.noSpaces}"
-              response <- http
-                            .request(TelluricSearchQuery)
-                            .withInput(input)
-              _        <- debug"raw GraphQL response: $response"
-              result   <- (response.data.map(_.search), response.errors) match
-                            case (Some(data), _)      =>
-                              data.pure[F]
-                            case (None, Some(errors)) =>
-                              for {
-                                _                 <-
-                                  warn"Telluric service errors: ${errors.toList.mkString("\n")}"
-                                // FIXME upstream
-                                isEmptyResultError =
-                                  errors.toList.exists(
-                                    _.message
-                                      .contains("Empty table cannot have column set to scalar value")
-                                  )
-                                result            <-
-                                  if (isEmptyResultError) {
-                                    warn"No telluric candidates found (server returned empty table error)"
-                                      .as(List.empty[TelluricStar])
-                                  } else {
-                                    error"GraphQL errors: $errors" *>
-                                      ApplicativeThrow[F].raiseError[List[TelluricStar]](
-                                        new RuntimeException(s"GraphQL errors: $errors")
-                                      )
-                                  }
-                              } yield result
-                            case _                    =>
-                              error"No data and no errors in GraphQL response" *>
-                                ApplicativeThrow[F].raiseError[List[TelluricStar]](
-                                  new RuntimeException("No data and no errors in GraphQL response")
-                                )
-            } yield result
+            TelluricSearchQuery[F]
+              .query(
+                input.coordinates.ra.toAngle.toDoubleDegrees.toFloat,
+                input.coordinates.dec.toAngle.toSignedDoubleDegrees.toFloat,
+                input.duration.toHours.toDouble.toFloat,
+                input.brightest.toFloat,
+                input.spType match
+                  case TelluricType.Hot               => "hot"
+                  case TelluricType.A0V               => "A0V"
+                  case TelluricType.Solar             => "Solar"
+                  case TelluricType.Manual(starTypes) =>
+                    starTypes.toList.mkString(",")
+              )
+              .raiseGraphQLErrors
+              .map(_.search)
 
   def noop[F[_]: Applicative]: TelluricTargetsClient[F] = new TelluricTargetsClient[F]:
     def search(input: TelluricSearchInput): F[List[TelluricStar]] =
